pyproject.toml
[project]
name = "markdown-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.1.0",
    "langchain-google-genai>=3.2.0",
    "langchain-postgres>=0.0.16",
    "langchain-text-splitters>=1.0.0",
    "mcp>=1.22.0",
    "pydantic-settings>=2.12.0",
]

[project.scripts]
markdown-rag = "markdown_rag:main"

[build-system]
requires = ["uv_build>=0.9.5,<0.10.0"]
build-backend = "uv_build"

[tool.ruff]
line-length = 80
target-version = "py311"

[tool.ruff.lint]
ignore = ["A002", "ANN003", "ANN204", "ANN401", "D100", "D417"]
select = ["A", "ANN", "D", "E", "F", "I", "N", "UP", "W"]

[tool.ruff.lint.flake8-annotations]
allow-star-arg-any = true
mypy-init-return = true

[tool.ruff.lint.pydocstyle]
convention = "google"
ignore-decorators = ["typing.overload"]

[tool.mypy]
plugins = ["pydantic.mypy"]
ignore_missing_imports = true
check_untyped_defs = true
follow_imports = "silent"
warn_return_any = true

[dependency-groups]
dev = [
    "mypy>=1.18.2",
]
